// web/src/generated-client/primeApi.schemas.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Prime Check Service
 * OpenAPI spec version: 0.0.0
 */
export interface Error {
  code: number;
  message: string;
}

export interface PrimeCheck {
  id: number;
  number: string;
  created_at: string;
  trace_id?: string;
  message_id?: string;
  is_prime?: boolean;
  status?: string;
}

export interface PrimeCheckInput {
  number: string;
}

export interface PrimeCheckList {
  items: PrimeCheck[];
}

export interface Setting {
  record_number_success: boolean;
  prime_check_success: boolean;
  email_send_success: boolean;
  dlq_save_success: boolean;
}